refactoring: test_page -> main
